move cavity_control
