+++ pyat/at/lattice/cavity_access.py
@@ -2,12 +2,52 @@
 from enum import Enum
 import numpy
 from typing import Optional
+import functools
 from .elements import RFCavity
 from .utils import AtError, make_copy, Refpts, get_bool_index
 from .lattice_object import Lattice
 
-__all__ = ['get_rf_frequency', 'get_rf_voltage', 'set_rf_voltage',
+__all__ = ['frequency_control',
+           'get_rf_frequency', 'get_rf_voltage', 'set_rf_voltage',
            'get_rf_timelag', 'set_rf_timelag', 'set_cavity', 'Frf']
+
+
+def frequency_control(func):
+    r""" Function to be used as decorator for
+    :pycode:`func(ring, *args, **kwargs)`
+
+    If :pycode:`ring.is_6d` is :py:obj:`True` **and** *dp*, *dct* or *df*
+    is specified in *kwargs*, make a copy of *ring* with a modified
+    RF frequency, remove *dp*, *dct* or *df* from *kwargs* and call
+    *func* with the modified *ring*.
+
+    If :pycode:`ring.is_6d` is :py:obj:`False` **or** no *dp*, *dct* or
+    *df* is specified in *kwargs*, *func* is called unchanged.
+
+    Examples:
+
+        .. code-block:: python
+
+            @frequency_control
+            def func(ring, *args, dp=None, dct=None, **kwargs):
+                pass
+    """
+
+    @functools.wraps(func)
+    def wrapper(ring, *args, **kwargs):
+        if ring.is_6d:
+            momargs = {}
+            for key in ['dp', 'dct', 'df']:
+                v = kwargs.pop(key, None)
+                if v is not None:
+                    momargs[key] = v
+            if len(momargs) > 0:
+                frequency = ring.get_revolution_frequency(**momargs) \
+                            * ring.harmonic_number
+                ring = set_cavity(ring, Frequency=frequency, copy=True)
+        return func(ring, *args, **kwargs)
+
+    return wrapper
 
 
 class Frf(Enum):
